address review comments
1. Separate lease.go file to have
   lease interfaces
2. Have spc_lease.go for spc specific
   lease implementation.
